edit field harga pada tabel barang

+++ database/migrations/2016_05_03_023547_create_barangs_table.php
@@ -16,7 +16,7 @@ class CreateBarangsTable extends Migration
             $table->increments('id');
             $table->string('nama');
             $table->string('jenis');
-            $table->float('harga');
+            $table->double('harga', 12, 2);
             $table->bigInteger('stok');
             $table->string('satuan');
             $table->timestamps();
